Add files via upload
Enhancing Keyword Search efficiency in big data: A comparative analysis of algorithms and Techniques
Statement of problem
With the blast of information in the advanced world, finding what we want rapidly has become progressively testing. Keywords, which we use to look for data on the web, are at the core of this test. We're confronted with questions like: How would we make keywords look more effective, particularly in the huge ocean of large information? As though that weren't sufficient, the ascent of voice colleagues like Siri and Google Collaborator adds one more layer of intricacy. In this way, the issue we're handling is the way to make catchphrase look through work better, particularly in the period of large information and voice partners. It's tied in with understanding how individuals search, what they battle with, and tracking down ways of making the entire cycle smoother and more compelling. Eventually, by taking care of this issue, we can make it simpler for everybody to find the data they're searching for in the midst of the consistently developing expanse of computerized information.
Research Question
How Enhancing Keyword Search Efficiency can be increased through the help of Voice Assistant and Big Data?